refactor(cli): move factory root CLI modules into factory/cli/ subpackage
Move 10 CLI modules from src/factory/ root into factory/cli/, extract
shared _store_connect helpers, update agent_cmd and test imports/patches,
and remove the src/factory folder-size exemption (root now 5 files).

Closes #1959
