+++ src/factory/cli/voice_smoke.py
