+++ tests/test_cli_wizard_create.py
@@ -214,7 +214,7 @@ class TestCreate:
         self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
     ) -> None:
         """Choosing 'u' at the location prompt writes to the user agents dir."""
-        import factory.cli_agent_create as create_mod
+        import factory.cli.agent_create as create_mod
 
         user_dir = tmp_path / "user_agents"
         monkeypatch.setattr(create_mod, "_user_agents_dir", lambda: user_dir)
@@ -246,7 +246,7 @@ class TestCreate:
         self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
     ) -> None:
         """Choosing 's' at the location prompt writes to the system agents dir."""
-        import factory.cli_agent_create as create_mod
+        import factory.cli.agent_create as create_mod
 
         system_dir = tmp_path / "system_agents"
         _user_dir = tmp_path / "user_agents"
