+++ tools/folder_exemptions.txt
@@ -10,5 +10,5 @@
 # factory_data_dir → $ROXABI_FACTORY_DIR). Top-level is correct for a
 # broadly-imported zero-dep resolver; cli_* decomposition (the real bloat)
 # is the follow-up. Plan: artifacts/plans/lyra-to-roxabi-factory-rename.md
-src/factory  # 15 files — +cli_secrets.py, secrets_reset.py (#1945 disaster recovery); cli_* split = follow-up
+
 src/factory/adapters/shared  # 16 files — glue kit extraction (#1931)
